Created maximum timeout for waiting for commands
The 'until' loops could cause a problem if there's an issue with
accessing the log file or database. Using a variable sleep with a
threshold, we can check a small and finite number of times with delays
of 2, 4, 8, 16 and 32 seconds. If we're not done by 32 seconds, don't
wait any longer.